requirements.txt
langgraph
langchain
langsmith
python-dotenv
langchain-groq
langchain-tavily

# for mcp project
langchain-mcp-adapters
mcp
langgraph

# langgraphstudio
langgraph-cli[inmem]